Refactored specs
